code/requirements.txt
python-dotenv==1.0.0
aiohttp==3.8.4

streamlit==1.23.1
streamlit_chat==0.0.2.2

# langchain
langchain==0.0.200
atlassian-python-api==3.38.0
beautifulsoup4==4.12.2
pytesseract==0.3.10
Pillow==9.5.0
lxml==4.9.2

# azure
azure-storage-blob==12.16.0
azure-ai-formrecognizer==3.2.1

# openai
openai==0.27.8

# Database
pymysql==1.0.3
redis==4.5.5